samples/01-dotenv-file/requirements.txt
bella-baxter>=1.0.0
python-dotenv>=1.0.0
